Add files via upload
According to the International Diabetes Federation (IDF) in 2022, the incidence of diabetes mellitus in Indonesia is quite high. It is reported that 463 million adults in the world suffer from diabetes with a global prevalence of 9.3%. However, the alarming condition is that 50.1% of diabetics are undiagnosed, making diabetes a silent killer that still haunts the world. It is estimated that the number of people with diabetes will increase by 45% to 629 million by 2045. In fact, 75% of people with diabetes in 2020 were aged 20-64 years (IDF, 2022).

The process of this project includes cleaning the data to remove missing values and normalization, and applying KNN for data classification. Prediction accuracy is evaluated using relevant metrics to measure the accuracy of the model.